Backend/requirements1.txt
asgiref==3.8.1
attrs==24.2.0
autobahn==24.4.2
Automat==24.8.1
bcrypt==4.2.0
certifi==2024.7.4
cffi==1.17.0
channels==4.1.0
charset-normalizer==3.3.2
constantly==23.10.4
cryptography==43.0.0
daphne==4.1.2
Django==4.2.15
django-cors-headers==4.4.0
django-prometheus==2.3.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
hyperlink==21.0.0
idna==3.7
incremental==24.7.2
passlib==1.7.4
pillow==10.4.0
prometheus_client==0.20.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
PyJWT==2.9.0
pyOpenSSL==24.2.1
python-dotenv==1.0.1
requests==2.32.3
service-identity==24.1.0
sqlparse==0.5.1
tomli==2.0.1
Twisted==24.7.0
txaio==23.1.1
typing_extensions==4.12.2
urllib3==2.2.2
zope.interface==7.0.1
psycopg2-binary
django-extensions==3.2.3
django-redis
channels-redis